# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Argus Alpha API
  version: 1.0.0-alpha
  description: Contract-first Alpha API for chat-first investing idea validation.
servers:
  - url: /api/v1
paths:
...
  /api/v1/backtests/run:
    post:
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Stable client key for deduplicating direct backtest runs.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestRunRequest"
      responses:
        "200":
          description: Completed Alpha backtest run.
          content:
            application/json:
              schema:
                type: object
                required: [run]
                properties:
                  run:
                    $ref: "#/components/schemas/BacktestRun"
        "400":
          description: Missing required Idempotency-Key header.
        "422":
          description: Problem Details for unsupported Alpha requests or unusable common data coverage. Coverage rejection occurs before quota admission.
        "409":
          description: The data window changed after approval and must be reviewed again.
        "503":
          description: >
            Problem Details when bounded auth-session verification is temporarily
            unavailable (`auth_session_verification_unavailable`) or market data
            is temporarily unavailable (`market_data_unavailable`).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/backtest-jobs/{id}:
    get:
      responses:
        "200":
          description: Durable async backtest job state, with linked run when available.
        "404":
          description: Job not found or inaccessible.
        "503":
          $ref: "#/components/responses/AuthSessionVerificationUnavailable"
  /api/v1/backtests/{id}:
    get:
      responses:
        "200":
          description: Backtest run detail including chart and trades.
        "503":
          $ref: "#/components/responses/AuthSessionVerificationUnavailable"
...
components:
  responses:
    AuthSessionVerificationUnavailable:
      description: Bounded Supabase session verification is temporarily unavailable; retry without redirecting to login.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: https://api.argus.app/problems/auth-session-verification-unavailable
            title: Session Verification Unavailable
            status: 503
            detail: Argus could not verify this session. Please try again.
            code: auth_session_verification_unavailable
            request_id: req_123
  schemas:
...
    BacktestRunRequest:
      type: object
      properties:
        strategy_id:
          type: string
        conversation_id:
          type: string
        template:
          type: string
        asset_class:
          enum: [equity, crypto, currency_pair]
        symbols:
          type: array
          minItems: 1
          maxItems: 5
          items:
            type: string
        timeframe:
          enum: [1h, 2h, 4h, 6h, 12h, 1D]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        side:
          type: string
        starting_capital:
          type: number
        allocation_method:
          const: equal_weight
        parameters:
          type: object
          additionalProperties: true
        benchmark_symbol:
          type: string
    BacktestRun:
      type: object
      required:
        - id
        - status
        - asset_class
        - symbols
        - benchmark_symbol
        - metrics
        - config_snapshot
        - conversation_result_card
      properties:
        id:
          type: string
          format: uuid
        status:
          enum: [queued, running, completed, failed]
        asset_class:
          enum: [equity, crypto]
        symbols:
          type: array
          maxItems: 5
          items:
            type: string
        allocation_method:
          const: equal_weight
        benchmark_symbol:
          type: string
          description: Equity defaults to SPY; crypto defaults to BTC.
        metrics:
          type: object
          description: Canonical aggregate and by-symbol backtest metrics.
        config_snapshot:
          type: object
          description: Reproducible normalized runtime configuration. New runs include requested_date_range, effective_date_range, and data_coverage; execution and display use the effective range while the requested range remains provenance.
        conversation_result_card:
          type: object
          description: Beginner-friendly inline chat result card. Completed chat runs may include P1 sidecar ids, decision state, and optional structured execution_costs evidence when the engine modeled fees/slippage.
...
    Error:
      type: object
      description: RFC 9457 Problem Details with code and request_id.
...